Step 9-3-B: harden adaptive learning state types and preserve fractional learning

+++ app_User_fb_C_feedforward_learning.h
@@ -13,14 +13,20 @@ typedef struct
 {
     int32_t error_threshold;
     int32_t gain;
-    int32_t counter;
+    uint32_t counter;
     int32_t offset;
     int32_t pid_sum;
     uint16_t pid_count;
     uint16_t stable_counter;
     uint16_t freeze_counter;
-    int16_t previous_sv;
+    int32_t previous_sv;
     APP_FB_BOOL sv_initialized;
+
+    /*
+     * Fractional learning remainder in Q15 offset units.
+     * This state intentionally survives normal valid learning windows so
+     * sub-one-count corrections can accumulate over time.
+     */
     int64_t learn_accumulator;
 } APP_FB_FF_LEARNING_T;
